requirements.txt
streamlit
numpy
firebase-admin
Pyrebase4
python-jwt
requests